Bitrix 16.5.5

+++ modules/catalog/general/catalog_export.php
@@ -14,32 +14,33 @@ class CAllCatalogExport
 {
 	
 	/**
-	* <p>Метод служит для проверки параметров, переданных в методы <a href="http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/add.php">CCatalogExport::Add</a> и <a href="http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/update.php">CCatalogExport::Update</a>. Метод динамичный.</p>
+	* <p>Метод служит для проверки параметров, переданных в методы <a href="http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/add.php">CCatalogExport::Add</a> и <a href="http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/update.php">CCatalogExport::Update</a>. Нестатический метод.</p>
 	*
 	*
-	* @param string $ACTION  Указывает, для какого метода идет проверка. Возможные значения:
-	* <br><ul> <li> <b>ADD</b> - для метода <a
-	* href="http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/add.php">CCatalogExport::Add</a>;</li> <li>
+	* @param string $ACTION  Указывает, для какого метода идет проверка. Возможные значения: 
+	* 			<br><ul> <li> <b>ADD</b> - для метода <a
+	* href="http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/add.php">CCatalogExport::Add</a>;</li> 				<li>
 	* <b>UPDATE</b> - для метода <a
-	* href="http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/update.php">CCatalogExport::Update</a>.</li> </ul>
+	* href="http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/update.php">CCatalogExport::Update</a>.</li>
+	* 			</ul>
 	*
 	* @param array &$arFields  Ассоциативный массив параметров профиля экспорта. Допустимые
-	* ключи: <ul> <li> <b>CREATED_BY</b> - ID пользователя, создавшего профиль;</li> <li>
-	* <b>MODIFIED_BY</b> - ID пользователя, изменившего профиль;</li> <li> <b>TIMESTAMP_X</b> -
-	* время последнего изменения профиля в формате сайта;</li> <li>
-	* <b>DATE_CREATE</b> - дата создания профиля в формате сайта;</li> <li> <b>FILE_NAME</b> -
-	* имя файла профиля со скриптом, осуществляющего экспорт. Ключ
-	* является обязательным, если $ACTION = 'ADD';</li> <li> <b>NAME</b> - название
-	* профиля экспорта. Ключ является обязательным, если $ACTION = 'ADD';</li> <li>
-	* <b>IN_MENU</b> - [Y|N] флаг отображения профиля в административном меню;</li>
-	* <li> <b>DEFAULT_PROFILE</b> - [Y|N] признак использования профиля по
-	* умолчанию;</li> <li> <b>IN_AGENT</b> - [Y|N] флаг наличия агента,
-	* осуществляющего автоматическое выполнение профиля экспорта; </li>
-	* <li> <b>IN_CRON</b> - [Y|N] флаг привязки профиля к утилите <i>cron</i> для
-	* автоматической периодической выгрузки (только для Unix-систем);</li>
-	* <li> <b>NEED_EDIT</b> - [Y|N] флаг означает неполную настройку профиля (до тех
-	* пор, пока профиль не будет отредактирован, он выполняться не
-	* будет). </li> </ul>
+	* ключи: 	         <ul> <li> <b>CREATED_BY</b> - ID пользователя, создавшего
+	* профиль;</li> <li> <b>MODIFIED_BY</b> - ID пользователя, изменившего профиль;</li>
+	* <li> <b>TIMESTAMP_X</b> - время последнего изменения профиля в формате
+	* сайта;</li> <li> <b>DATE_CREATE</b> - дата создания профиля в формате сайта;</li>
+	* <li> <b>FILE_NAME</b> - имя файла профиля со скриптом, осуществляющего
+	* экспорт. Ключ является обязательным, если $ACTION = 'ADD';</li> <li> <b>NAME</b> -
+	* название профиля экспорта. Ключ является обязательным, если $ACTION =
+	* 'ADD';</li> <li> <b>IN_MENU</b> - [Y|N] флаг отображения профиля в
+	* административном меню;</li> <li> <b>DEFAULT_PROFILE</b> - [Y|N] признак
+	* использования профиля по умолчанию;</li> <li> <b>IN_AGENT</b> - [Y|N] флаг
+	* наличия агента, осуществляющего автоматическое выполнение
+	* профиля экспорта; </li> <li> <b>IN_CRON</b> - [Y|N] флаг привязки профиля к
+	* утилите <i>cron</i> для автоматической периодической выгрузки (только
+	* для Unix-систем);</li> <li> <b>NEED_EDIT</b> - [Y|N] флаг означает неполную
+	* настройку профиля (до тех пор, пока профиль не будет
+	* отредактирован, он выполняться не будет). </li> </ul>
 	*
 	* @return bool <p>В случае корректности переданных параметров возвращает <i>true</i>,
 	* иначе - <i>false</i>.</p>
@@ -141,13 +142,13 @@ class CAllCatalogExport
 
 	
 	/**
-	* <p>Метод удаляет профиль экспорта с кодом ID. Метод динамичный.</p>
+	* <p>Метод удаляет профиль экспорта с кодом ID. Нестатический метод.</p>
 	*
 	*
-	* @param int $ID  Код удаляемого профиля экспорта.
+	* @param mixed $intID  Код удаляемого профиля экспорта.
 	*
 	* @return bool <p>Возвращает <i>true</i> в случае успешного удаления и <i>false</i> - в
-	* противном случае.</p> <br><br>
+	* противном случае.</p><br><br>
 	*
 	* @static
 	* @link http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/delete.php
@@ -157,19 +158,21 @@ class CAllCatalogExport
 	{
 		global $DB;
 
-		$ID = intval($ID);
+		$ID = (int)$ID;
+		if ($ID <= 0)
+			return false;
 		return $DB->Query("DELETE FROM b_catalog_export WHERE ID = ".$ID." AND IS_EXPORT = 'Y'", true);
 	}
 
 	
 	/**
-	* <p>Возвращает список профилей экспорта по фильтру <i>arFilter</i>, отсортированый в соответствии с <i>arOrder</i>. Метод динамичный.</p>
+	* <p>Возвращает список профилей экспорта по фильтру <i>arFilter</i>, отсортированый в соответствии с <i>arOrder</i>. Нестатический метод.</p>
 	*
 	*
 	* @param array $arOrder = array("ID"=>"ASC") Массив, в соответствии с которым сортируются результирующие
-	* записи. Массив имеет вид: <pre class="syntax">array( "название_поля1" =&gt;
+	* записи. Массив имеет вид: 		<pre class="syntax">array( "название_поля1" =&gt;
 	* "направление_сортировки1", "название_поля2" =&gt;
-	* "направление_сортировки2", . . . )</pre> В качестве "название_поля<i>N</i>"
+	* "направление_сортировки2", . . . )</pre> 		В качестве "название_поля<i>N</i>"
 	* могут использоваться: <ul> <li> <i>NAME</i> - название профиля экспорта;</li>
 	* <li> <i>FILE_NAME</i> - имя файла профиля со скриптом, осуществляющего
 	* экспорт;</li> <li> <i>DEFAULT_PROFILE</i> - Y|N] флаг использования профиля по
@@ -190,13 +193,13 @@ class CAllCatalogExport
 	* результат сортируется по второму и т.д.).
 	*
 	* @param array $arFilter = array() Массив, в соответствии с которым фильтруются записи профилей
-	* экспорта. Массив имеет вид: <pre class="syntax">array(
+	* экспорта. Массив имеет вид: 		<pre class="syntax">array(
 	* "[модификатор]название_поля1" =&gt; "значение1",
 	* "[модификатор]название_поля2" =&gt; "значение2", . . . )</pre>
 	* Удовлетворяющие фильтру записи возвращаются в результате, а
 	* записи, которые не удовлетворяют условиям фильтра,
-	* отбрасываются.<br><br> Допустимым является следующий модификатор:
-	* <ul> <li> <b> !</b> - отрицание;</li> </ul> В качестве "название_поляX" может
+	* отбрасываются.<br><br> 	Допустимым является следующий модификатор:
+	* 		<ul> <li> <b> 	!</b>  - отрицание;</li> 		</ul> В качестве "название_поляX" может
 	* стоять любое поле профиля экспорта, кроме <i>SETUP_VARS</i>, <i>TIMESTAMP_X</i> и
 	* <i>DATE_CREATE</i>.
 	*
@@ -205,12 +208,12 @@ class CAllCatalogExport
 	* Необязательный. По умолчанию равен <i>false</i>.
 	*
 	* @return CDBResult <p>Возвращается объект класса CDBResult, содержащий коллекцию
-	* ассоциативных массивов с ключами:</p> <ul> <li> <b>ID</b> - код записи;</li> <li>
+	* ассоциативных массивов с ключами:</p><ul> <li> <b>ID</b> - код записи;</li> <li>
 	* <b>FILE_NAME</b> - имя файла профиля со скриптом, осуществляющего
 	* экспорт;</li> <li> <b>NAME</b> - название профиля экспорта;</li> <li> <b>IN_MENU</b> -
 	* [Y|N] флаг отображения профиля в административном меню;</li> <li>
 	* <b>IN_AGENT</b> -[Y|N] флаг наличия агента, осуществляющего автоматическое
-	* выполнение профиля экспорта; </li> <li> <b>IN_CRON</b> - [Y|N] флаг привязки
+	* выполнение профиля экспорта;  </li> <li> <b>IN_CRON</b> - [Y|N] флаг привязки
 	* профиля к утилите <i>cron</i> для автоматической периодической
 	* выгрузки;</li> <li> <b>SETUP_VARS</b> - параметры настройки профиля в виде
 	* url-строки;</li> <li> <b>DEFAULT_PROFILE</b> - [Y|N] флаг использования профиля по
@@ -222,13 +225,13 @@ class CAllCatalogExport
 	* <b>CREATED_BY</b> - ID пользователя, создавшего профиль;</li> <li> <b>MODIFIED_BY</b> - ID
 	* пользователя, изменившего профиль;</li> <li> <b>TIMESTAMP_X</b> - дата и время
 	* последнего изменения профиля в формате сайта;</li> <li> <b>DATE_CREATE</b> -
-	* дата и время создания профиля в формате сайта.</li> </ul> <br><br>
+	* дата и время создания профиля в формате сайта.</li> </ul><br><br>
 	*
 	* @static
 	* @link http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/getlist.php
 	* @author Bitrix
 	*/
-	public static function GetList($arOrder=array("ID"=>"ASC"), $arFilter=array(), $bCount = false)
+	public static function GetList($arOrder = array("ID"=>"ASC"), $arFilter = array(), $bCount = false)
 	{
 		global $DB;
 		$arSqlSearch = array();
@@ -241,7 +244,8 @@ class CAllCatalogExport
 		for ($i = 0, $intCount = count($filter_keys); $i < $intCount; $i++)
 		{
 			$val = $DB->ForSql($arFilter[$filter_keys[$i]]);
-			if (strlen($val)<=0) continue;
+			if (strlen($val)<=0)
+				continue;
 
 			$bInvert = false;
 			$key = $filter_keys[$i];
@@ -365,18 +369,18 @@ class CAllCatalogExport
 
 	
 	/**
-	* <p>Метод возвращает информацию о профиле экспорта с заданным ID. Метод динамичный.</p>
+	* <p>Метод возвращает информацию о профиле экспорта с заданным ID. Нестатический метод.</p>
 	*
 	*
-	* @param int $ID  Код записи.
+	* @param mixed $intID  Код записи.
 	*
 	* @return array <p>Метод возвращает ассоциативный массив параметров профиля с
-	* ключами:</p> <ul> <li> <b>ID</b> - код записи;</li> <li> <b>FILE_NAME</b> - имя файла
+	* ключами:</p><ul> <li> <b>ID</b> - код записи;</li> <li> <b>FILE_NAME</b> - имя файла
 	* профиля со скриптом, осуществляющего экспорт;</li> <li> <b>NAME</b> -
 	* название профиля экспорта;</li> <li> <b>IN_MENU</b> - [Y|N] флаг отображения
 	* профиля в административном меню;</li> <li> <b>IN_AGENT</b> -[Y|N] флаг наличия
 	* агента, осуществляющего автоматическое выполнение профиля
-	* экспорта; </li> <li> <b>IN_CRON</b> - [Y|N] флаг привязки профиля к утилите
+	* экспорта;  </li> <li> <b>IN_CRON</b> - [Y|N] флаг привязки профиля к утилите
 	* <i>cron</i> для автоматической периодической выгрузки;</li> <li> <b>SETUP_VARS</b>
 	* - параметры настройки профиля в виде url-строки;</li> <li> <b>DEFAULT_PROFILE</b> -
 	* [Y|N] флаг использования профиля по умолчанию;</li> <li> <b>LAST_USE</b> - дата
@@ -388,7 +392,7 @@ class CAllCatalogExport
 	* создавшего профиль;</li> <li> <b>MODIFIED_BY</b> - ID пользователя, изменившего
 	* профиль;</li> <li> <b>TIMESTAMP_X</b> - дата и время последнего изменения
 	* профиля в формате сайта;</li> <li> <b>DATE_CREATE</b> - дата и время создания
-	* профиля в формате сайта.</li> </ul> <br><br>
+	* профиля в формате сайта.</li> </ul><br><br>
 	*
 	* @static
 	* @link http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/getbyid.php
@@ -417,14 +421,14 @@ class CAllCatalogExport
 
 	
 	/**
-	* <p>Метод выполняет профиль <i>profile_id</i> на агенте. Метод динамичный.</p>
+	* <p>Метод выполняет профиль <i>profile_id</i> на агенте. Нестатический метод.</p>
 	*
 	*
 	* @param int $profile_id  Код выполняемого профиля.
 	*
 	* @return mixed <p>В случае успешного выполнения профиля экспорта метод
 	* возвращает строку для следующего вызова агента. В противном
-	* случае метод вернет <i>false</i>.</p> <br><br>
+	* случае метод вернет <i>false</i>.</p><br><br>
 	*
 	* @static
 	* @link http://dev.1c-bitrix.ru/api_help/catalog/classes/ccatalogexport/pregenerateexport.php
@@ -435,7 +439,8 @@ class CAllCatalogExport
 		global $DB;
 
 		$profile_id = (int)$profile_id;
-		if ($profile_id <= 0) return false;
+		if ($profile_id <= 0)
+			return false;
 
 		$ar_profile = CCatalogExport::GetByID($profile_id);
 		if ((!$ar_profile) || ('Y' == $ar_profile['NEED_EDIT']))
@@ -446,9 +451,7 @@ class CAllCatalogExport
 		{
 			$strFile = CATALOG_PATH2EXPORTS_DEF.$ar_profile["FILE_NAME"]."_run.php";
 			if (!file_exists($_SERVER["DOCUMENT_ROOT"].$strFile))
-			{
 				return false;
-			}
 		}
 
 		$arSetupVars = array();
@@ -457,9 +460,7 @@ class CAllCatalogExport
 		{
 			parse_str($ar_profile["SETUP_VARS"], $arSetupVars);
 			if (!empty($arSetupVars) && is_array($arSetupVars))
-			{
 				$intSetupVarsCount = extract($arSetupVars, EXTR_SKIP);
-			}
 		}
 
 		global $arCatalogAvailProdFields;
@@ -497,6 +498,4 @@ class CAllCatalogExport
 
 		return "CCatalogExport::PreGenerateExport(".$profile_id.");";
 	}
-
 }
-?>
